Document the HexAddress coordinate systems
HexAddress values can be defined according to three different coordinate
systems. While this is not necessarily ideal, the documentation now
defines each of these systems and provides code examples.

The HexAddress type now supports negative row and column numbers, which
allows us to place tiles adjacent to any map hex (as opposed to having
fixed boundaries at row = 0 and at column = 0). So we can define "A1" to
be equal to the origin (0, 0), rather than supporting a game-specific
"offset" to adjust between string coordinates and numeric coordinates.

+++ examples/tile_catalogue.rs
@@ -41,8 +41,8 @@ fn place_tiles(
 
     // Build an iterator over the map hexes.
     let mut tile_addrs: Vec<String> = vec![];
-    let rows: Vec<usize> = (0..rows).collect();
-    let cols: Vec<usize> = (0..cols).collect();
+    let rows: Vec<isize> = (0..rows as isize).collect();
+    let cols: Vec<isize> = (0..cols as isize).collect();
     for row in &rows {
         for col in &cols {
             let addr = HexAddress::new(*row, *col);
